+++ src/schemathesis/exceptions.py
@@ -1,9 +1,42 @@
-from typing import Union
+from hashlib import sha1
+from typing import Dict, Type, Union
 
 import attr
 import requests
+from jsonschema import ValidationError
 
 from .utils import WSGIResponse
+
+CACHE: Dict[Union[str, int], Type[AssertionError]] = {}
+
+
+def get_exception(name: str) -> Type[AssertionError]:
+    """Create a new exception with provided name or fetch one from cache."""
+    if name in CACHE:
+        exception_class = CACHE[name]
+    else:
+        exception_class = type(name, (AssertionError,), {})
+        CACHE[name] = exception_class
+    return exception_class
+
+
+def get_status_code_error(status_code: int) -> Type[AssertionError]:
+    """Return new exception for an unexpected status code."""
+    name = f"StatusCodeError{status_code}"
+    return get_exception(name)
+
+
+def get_response_type_error(expected: str, received: str) -> Type[AssertionError]:
+    """Return new exception for an unexpected response type."""
+    name = f"SchemaValidationError{expected}_{received}"
+    return get_exception(name)
+
+
+def get_schema_validation_error(exception: ValidationError) -> Type[AssertionError]:
+    """Return new exception for schema validation error."""
+    message_digest = sha1(str(exception).encode("utf-8")).hexdigest()
+    name = f"SchemaValidationError{message_digest}"
+    return get_exception(name)
 
 
 class InvalidSchema(Exception):